[Initial Commit] - Add the code with instructions
The code necessary for the contribution is added. Although much of the code was removed and changed to remove "company secrets". Some functionality has also been deprecated, as it was specific to the former repository and project.

The Unit test has also been changed.

The unit test, along with the base code must be edited so that it works.